requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.2.0
PyMuPDF==1.23.19
PyMuPDFb==1.23.9
pypng==0.20220715.0
qrcode==7.4.2
typing_extensions==4.9.0
Werkzeug==3.0.3
